camera rotating work
camera is rotating properly, but some errors are still made. Sometimes when rotating, the player would move +32 in Y  or +64 in X. Still resulting in the same position.
Solution is to draw it on paper and see what offsets are correct after rotating, since dimensions change, offset for X or Y can be max - X, max - Y added or subtracted...